// types.ts
export enum Network {
  AVALANCHE = "Avalanche",
  ARBITRUM = "Arbitrum",
  BSC = "BSC",
  ETHEREUM = "Ethereum",
  FANTOM = "Fantom",
  OPTIMISM = "Optimism",
  POLYGON = "Polygon",
}

export enum Token {
  STG = "STG",
}

export enum Native {
  AVAX = "AVAX",
  BNB = "BNB",
  ETH = "ETH",
  FTM = "FTM",
  MATIC = "MATIC",
}